app/requirements.txt
fastapi
uvicorn
psycopg2
python-dotenv